Move kas_core::event::ErasedMessage -> kas_core::Erased
